BIGGGGG map format change (crystal)

+++ Server/Envir/SConnection.cs
@@ -735,7 +735,6 @@ namespace Server.Envir
         {
             if (Stage != GameStage.Game && Stage != GameStage.Observer) return;
 
-
             S.MarketPlaceHistory result = new S.MarketPlaceHistory { Index = p.Index, Display = p.Display, ObserverPacket = false };
             Enqueue(result);
 
@@ -747,7 +746,7 @@ namespace Server.Envir
             result.LastPrice = info.LastPrice;
 
             long average = 0;
-            int count = 0;
+            long count = 0;
 
             foreach (long value in info.Average)
             {
